edusprint-/requirements.txt
Django>=5.2.3
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
mysqlclient>=2.2.0
pymysql>=1.1.0
cryptography>=41.0.0
Pillow>=10.0.0
python-decouple>=3.8
django-cors-headers>=4.3.0 